Don't change routine names.
There was a bug in this code that caused the "()" not to be included
after the function name in the create script.

Also, I disagree with the concept SchemaZen making changes to the
schema. I can see the convienence of it, but I think it's SchemaZen
should stick to dumping and re-creating the schema as is. Leave the job
of fixing things to some other tool.