Remove ScriptCreate() method from Database
This was only used in tests and has different behavior with respect to
resolving dependencies between db objects than ScriptToDir / CreateFromDir.

I'm removing it to make sure tests use CreateFromDir and verify the path
acutally used by the cli.

This is in preperation to fix the issue with users assigned to role #188

Also added two failings test for the user / role bug.