Fix circular dependency issue
fixes #188

This also changed the behavior of the script command. Schemas are now
written to separate scripts in the schemas directory.

The create command remains backward compatible because prior ot this
change schemas were written to a single script in the schemas dir. The
command will work with either a single script or multiple.